Workaround for a clean remove.
Added a routine which checks if the file
/var/lib/opkg/info/enigma2-plugin-extensions-backupsuite.list contains
the .pyo files and the file speed.txt, if not it is added (normal on
first run). After these files are added the backupsuite leaves no
remains when it is uninstalled. This workaround is necessary because the
.pyo's are ignored when build from the github. On install however these
are created but the package is unaware of these files so the pyo's left
behind after an uninstall.
This could be corrected with an postrm routine but this isn't added in
the recpe.
A bit crude workaround but it's better then no workaround