Added some kind of support for unknown receivers.
If the receiver is not found in the lookuptable the program won't be
aborted but a backup will be made with some standard file naming. The
user has at least to rename the folder where the backup is made. Highly
experimental but it is worth a go. WTE gave me the idea to read some
necessities from the receiver. Thanks!