Added backupfile management by Tension and debugged the Solo4k handling
Some big changes, at first now it is possible to maintain your backups
from the plugin.
With the blue button you are now able to delete the
backups you want to get rid of.
Please bear in mind that this only works
for the backups created with this or any newer version of the
BackupSuite.
If you want to be able to do this also for your older
backups then you'll have to move the files:
imageversion from the data
folder to the folder above (in the folder right under the folder
/fullbackup_<model/<date_time>/ )
Big thanks to Tension for the
changes

Also debugged a bit for the Solo4k, now you should be able to
make a backup from the solo4k. There is still some work to do but for
now it should work.
Please remember: the solo4k backup can't be restored
by the BackupSuite using the ofgwrite utility (yet).
Thanks Findex for
doing some testing with his Solo4k!