Reworked the backup routine for the Mutant HD51
Reworked the routine for the Mutant HD51. In the former version the
possibility existed that the made backup was not complete.
Now the
backup is made of the whole partition and then it's compressed in
rootfs.tar.bz2 format.

The backup can ONLY be restored using the
BackupSuite -> Blue button (=Restore backup).
It can't be flashed by
the USB-stick.

Thanks go to Betacentauri which pointed me to the
possible problems and gave me some guidance to the solution!
It's still
work in progress so it could be that in (short) term there will be more
changes.
Please test and report back if there are any issues (or if
everthing goes like intended!)